Mostly-complete auto-installation of the Android SDK and components.
If the (broken) global config property is selected, the SDK will be downloaded
and installed as part of any build using this plugin, assuming an existing SDK
isn't already installed.

Currently this works (on a Mac), but should be fine on other platforms.
Support for slave nodes and HTTP(S) proxies is still missing.