Prevent multiple instances of the same emulator from starting concurrently.
Implemented a QueueTaskDispatcher which determines the requested emulator
configuration and checks whether any executors on a given machine are
currently executing that configuration; if so the new build is rejected.

This increases the required Jenkins/Hudson version from 1.350 to 1.360.

Thanks to Kohsuke for his assistance at the Jenkins Hackathon in Hamburg!