[JENKINS-11973] Option for emulators in $WORKSPACE
This adds a global option that essentially does
ANDROID_SDK_HOME=$WORKSPACE, causing emulators to be put into
$WORKSPACE/.android.

This allows you to run concurrent jobs on the same slave, and to run the
same emulator config from two different jobs on the same slave. It also
makes the emulators more isolated and allows you to clean-slate the
emulators by wiping out the workspace.