Prevent dependent platforms from being installed redundantly.
Previously, when installing an SDK add-on, both the add-on and the
dependent Android platform version would be installed, regardless
of whether the dependent platform was already installed - Android's
installer tool does as its told and simply downloads it again.

Now, we do an extra check on each component of a platform before
installing, and ignore any components that are already installed.