'isUnix' flag mostly unneeded.
In relation to a531f90453e25378efc5a1b1e15b55a99c25f709, I noticed that
the way the 'isUnix' flag is used is mostly redundant. It's computed by
the master and then passed with a closure to be exected on the slave.

But once your code starts to execute on the slave, its JVM already knows
what platform it is running on. So it's more concise and faster not to
pass around this flag from the master.

Similarly, when a method signature returns a File, like
Utils.getHomeDirectory(), then you are by definition already executing
on the target JVM, so these functions shouldn't take the 'isUnix'
parameter. It's concise, faster, and less error prone to just let this
function computes whether it's running on an Unix or not.

This commit eliminates those redundant uses of the 'isUnix' flag from
the signature.

I don't know if other plugins started depending on these signatures,
and/or if you'd like to maintain API compatibility, so I'm sending this
in as a pull request. Please feel free to disregard if that is the case.