Improve the heuristic used to determine which app project a test project belongs to.
Previously, we just took the first app project found in alphabetical order.
This didn't work well, e.g. with the following project layout:
- actionbarsherlock/  (library project)
- app/                (app project)
- tests/              (test project)

In this case, the "tests" project would be paired with "actionbarsherlock/samples/demos".

Now we search all app projects and take the one "nearest" to the test project,
in terms of relative path distance between the two directories.