increase-emulator-start-time-limit: Increase limit from 180s to 360s
The emulator sometimes takes more than 180 seconds, the hard coded time
limit[1][2], to be ready, when starting from a snapshot.

[1]: https://github.com/jenkinsci/android-emulator-plugin/blob/9b91a1bb1c99be0b5371f0ada2992bd6ed51d503/src/main/java/hudson/plugins/android_emulator/AndroidEmulator.java#L63
[2]: https://github.com/jenkinsci/android-emulator-plugin/blob/9b91a1bb1c99be0b5371f0ada2992bd6ed51d503/src/main/java/hudson/plugins/android_emulator/AndroidEmulator.java#L381

The plugin increases the time limit by a factor of 4 when it's not using
a snapshot.

I saw the non-snapshot scenario take between 400 and 500 seconds to get
the emulator ready.

I increased the time limit to 360 s, and reduced the factor from 4 to 2.