[FIXED JENKINS-17532] Ensure system images are installed as required.
Previously, if the platform itself was installed — e.g. due to using the
"Install Android project prerequisites" build step — we wouldn't check, when
trying to start an emulator, whether any required system images were also
installed. If they were missing, emulator creation would fail every time.

Now we check that both the platform and at least one system image have been
installed for the desired platform.

If there are no system images installed, all images will be downloaded.
This is slightly dumb at the moment, as we will also re-download the platform,
even if it already exists, and we won't download any system images if there is
at least one installed for the given platform — regardless of whether it's the
one required for the configured emulator.