[FIXED JENKINS-11083] Allow adding a suffix to generated AVD names.
This makes it possible to use the exact same emulator config in two jobs without
one job having to block waiting for the other job to finish using the emulator.

This patch comes from pull requests #9 and #33, with some additional changes to
documentation and assuring that generated AVD names are valid, despite the
suffix field being free text.