Wait for package manager to be available before trying to install APKs.
Previously, even although the system appears to have already started, when
attempting to install an APK, it could fail with an error message saying
"Could not access the Package Manager", which would then fail the build.

A similar issue was reported for Android Studio (Android bug #62113), which was
resolved by additionally waiting for the "android.process.acore" process.

The plugin now does the same, with a two-minute timeout, as, on a heavily-loaded
system, I observed emulators waiting approximately a minute for that process.