Improve error logging when AVD creation fails.
This will help to diagnose problems such as those found in
JENKINS-36885 however it is not a "fix" for the problem which
is more likely to be a job configuration error.

Specifically this patch always logs any stdout or stderr
output if stderr is produced and the error is not a
specifically handled case.

In addition the attempt to set the hw.lcd.density property
is delayed until the output has been generated so that it
does not prevent the logging of any diagnostic information.