Update system image component name construction for APIs >= 25
The naming convention for Android system images changes slightly beginning
at level 25; this updates the component name constructor logic so that the
desired image is found.

N.B. Based on the images available via the Device Manager GUI and on
searching around the web, it looks like only system images with Google API
support are available for these API levels.

Bug: https://phabricator.wikimedia.org/T150623
Change-Id: I9c42dfc0d8ee662440fcdd1668007d0931a7b2b3