[Test] Stop forcing QEMU 1 use with the "-engine classic" flag
For builds in which the Android SDK tools are at major version 25 or
higher, the upstream Jenkins Android emulator plugin adds the command-
line option "-engine classic"[1] when creating the Android virtual device,
in order to force QEMU 1 use. Unfortunately, this is hard-coded to
happen based on the build environment rather than being exposed as a user-
configurable option. This was apparently intended to work around a bug in
which the -ports option was unavailable when using QEMU 2, though that bug
has reportedly since been fixed.[2]

The plugin unconditionally creates the AVD with snapshot support but the
emulator would then force select the classic engine. Only add snapshot
support (android created avd -a) when the job requests it.

I believe the "-engine classic" option may be behind our boot timeout
issues as that trouble started with the previous emulator around the time
we bumped our SDK tooling to level 25. This patch is intended to test that
hypothesis.

[1] https://developer.android.com/studio/run/emulator-commandline.html
[2] https://issues.jenkins-ci.org/browse/JENKINS-40178

Change-Id: I0e46f14f0986f9982cdf17b65820b6cd79423477