[FIXED JENKINS-40178][FIXED JENKINS-43215][FIXED JENKINS-44490] Add support for Android Emulator 2.0
Added support for the 'new' Android Emulator 2.0, which got support in SDK Tools v25 but
was lacking some functionality the plugin relies on. Since SDK Tools v26 the new emulator
engine supports all needed flags and parameters.

The plugin creates AVDs and starts the emulator in legacy format or in the new format,
dependent of which SDK Tools version is installed (<= 25 legacy, >= 26 new format).

A new configuration option was added ('Device definition') which is needed (to avoid
user input) on AVD creation. It is used to define a specific device ('avdmanager list device'),
but it is only respected if SDK Tools >= 26 are used.

Added a helper class AndroidPlatformVersions which holds the current data of platforms,
which was moved from AndroidPlatform. Additional AndroidPlatform got refactored to
handle and retrieve more information needed for the emulator (package path) and
for more convenient access to vendor and api information of the used platform.