Fix system-image-path if custom tag is used in ABI
Having a tag prefixed to the ABI-String produced an
incorrect system-image value for the sdkmanager to
load and an incorrect package path for the avdmanager
to create the AVD.
Both occurences are fixed, additionally the description
of the tragetABI field was updated, to give more clear
instruction about how to prefix a tag.
Always specify --tags on create avd, autoselection does
not work on SDK Tools v26.0.1 and lower.